Fix Linux and HOL Install Scripts

+++ FULLTEXT-LITE/install/FULLTEXT-LITE.sh
@@ -389,5 +389,8 @@ SERVER=${5}
 demohome="$(dirname "$(pwd)")"
 logfilename=$demohome/install/FULLTEXT-LITE.log
 echo "Log File : $logfilename"
-rm $logfilename
+if [ -f "$logfilename" ]
+then
+  rm $logfilename
+fi
 doInstall 2>&1 | tee -a $logfilename
