+++ XQUERY/install/XQUERY.sh
@@ -514,5 +514,8 @@ SERVER=${5}
 demohome="$(dirname "$(pwd)")"
 logfilename=$demohome/install/XQUERY.log
 echo "Log File : $logfilename"
-rm $logfilename
+if [ -f "$logfilename" ]
+then
+  rm $logfilename
+fi
 doInstall 2>&1 | tee -a $logfilename
